agent/pyproject.toml
[project]
name = "a2ui-expense-reporter"
version = "0.1.0"
description = "A2UI expense reporting agent with OCR and local JSON storage."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "a2a-sdk[http-server]>=0.3.0",
  "a2ui @ file:///Users/arakawayasuaki/git/expense-ui/a2ui-extension",
  "click>=8.1.8",
  "python-dotenv>=1.1.0",
  "pytesseract>=0.3.10",
  "pdf2image>=1.17.0",
  "Pillow>=10.4.0",
  "starlette>=0.37.2",
  "sse-starlette>=2.1.0",
  "uvicorn>=0.30.6",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
